General code cleanup and addition of more operations
Separated the custom functions for the ExpressionBuilder library into their own files to reduce clutter, removed the functionality that kept the last operation done in favor of simply keeping the result of said operation. Lastly, added functionality to the writeExpression function that checks the string passed in and converts it to the appropriate equivalent before appending it to the expression string.